Add files via upload
In this notebook, we have already written a basic level that already recognizes masks in some way, but still does it quite poorly. First, the grid doesn't know the "no mask" class (this is visible on the laptop), this is because Fast R-CNN recognizes class 0 as the background, and also doesn't recognize the "wrong mask"class very well. There is also the problem that the image drawing feature violates the DRY principle. And the written function area_boxes, which calculates the IOU, is not quite complete yet.