version_2
Here, changes were made to the generate_boxes function, class 0 was removed due to the reasons that were described in the first version of the laptop, now the model recognizes the "no mask" class. The plot_image function has been improved, now there is no violation of the DRY principle.