final_version
This is the final version of the project. The model is trained for 100 epochs, using the SGD optimizer with nesterov=True. Upgraded the area_boxes function and added the choose function, which improves the selection of boxes in the image. The lowest losses were achieved, as well as improvements in the quality of recognition of all classes. To improve the recognition of the "incorrect mask" class, reducing the batch size helped. The results of the model are also shown in an image downloaded from the Internet. Added saves for our model and best weights.